Fix compile warnings in Windows build.

+++ airdrums/Source/MainComponent.cpp
@@ -424,8 +424,8 @@ void MainContentComponent::renderOpenGL()
         if (tempoControl)
             tempoControl->setBounds(HUDRect((GLfloat) Environment::instance().screenW / 2.f - tempoControlWidth / 2.f,
                                             (GLfloat) Environment::instance().screenH - 70 / 2.f - tempoControlHeight / 2.f + 5,
-                                            tempoControlWidth,
-                                            tempoControlHeight));
+                                            (GLfloat) tempoControlWidth,
+                                            (GLfloat) tempoControlHeight));
         
         layoutPadsLinear();
         sizeChanged = false;
